+++ utils/to_rad.c
@@ -6,13 +6,13 @@
 /*   By: vguttenb <vguttenb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/15 19:46:04 by vguttenb          #+#    #+#             */
-/*   Updated: 2022/08/15 19:46:12 by vguttenb         ###   ########.fr       */
+/*   Updated: 2023/01/26 21:31:53 by vguttenb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cube3d.h"
 
-float	to_rad(float grad)
+double	to_rad(double grad)
 {
 	return (grad * RADGR);
 }
